+++ Extra Profiles and Skill Sets/lua/menumanager.lua
@@ -1,15 +1,13 @@
 Hooks:PostHook(MenuManager, "init", "EPSS-PostHook-MenuManager:init", function(...)
-	local item = EPSS:get_menu_item("total_profiles_deferred")
+	--Slider minimum is either 1 or base game number of profiles.
+	EPSS:adjust_slider_min()
+
+	--Adjust slider max if someone manually edited their save
+	local item = ncUtils:get_menu_item("total_profiles", EPSS)
 	if item then
-		if not EPSS.settings.allow_fewer then
-			item._min = EPSS._base_num_profiles
-		else
-			item._min = 1
-		end
-		--Adjust slider max if someone manually edited their save
-		if EPSS.settings.total_profiles_deferred > item._max then
-			item._max = EPSS.settings.total_profiles_deferred
-			item._value = EPSS.settings.total_profiles_deferred
+		if EPSS._session_profiles > item._max then
+			item._max = EPSS._session_profiles
+			item._value = EPSS._session_profiles
 		end
 	end
 end)
@@ -17,20 +15,15 @@ end)
 --Not allowed to switch skill sets if not suspended
 local orig_MenuCallbackHandler_set_active_skill_switch = MenuCallbackHandler.set_active_skill_switch
 function MenuCallbackHandler:set_active_skill_switch(item)
-	if EPSS.settings.autobind_skills_2 then
-		local skill_idx = item:parameters().name
-		local profile_idx = managers.multi_profile and managers.multi_profile._global._current_profile
-		if profile_idx and profile_idx ~= skill_idx then
-			local switch_data = managers.skilltree and managers.skilltree._global.skill_switches[profile_idx]
-			if switch_data and switch_data.unlocked and not managers.skilltree:is_skill_switch_suspended(switch_data) then
-				managers.skilltree:switch_skills(profile_idx)
-				self:refresh_node()
-				local menu_title = managers.localization:text("epss_dialog_title")
-				local menu_message = managers.localization:text("epss_dialog_autobind")
-				local menu_options = {{text = managers.localization:text("dialog_continue"), is_cancel_button = true}}
-				QuickMenu:new(menu_title, menu_message, menu_options, true)
-				return
-			end
+	if EPSS.settings.autobind_skillsets then
+		local profile_idx = managers.multi_profile:current_profile_index()
+		local switch_data = managers.skilltree._global.skill_switches[profile_idx]
+		if switch_data and switch_data.unlocked and not managers.skilltree:is_skill_switch_suspended(switch_data) then
+			local menu_title = managers.localization:text("epss_dialog_title")
+			local menu_message = managers.localization:text("epss_dialog_skill_switch_blocked")
+			local menu_options = {{text = managers.localization:text("dialog_ok"), is_cancel_button = true}}
+			QuickMenu:new(menu_title, menu_message, menu_options, true)
+			return
 		end
 	end
 	orig_MenuCallbackHandler_set_active_skill_switch(self, item)
@@ -40,11 +33,8 @@ end
 local orig_MenuCallbackHandler_unsuspend_skill_switch_dialog_yes = MenuCallbackHandler.unsuspend_skill_switch_dialog_yes
 function MenuCallbackHandler:unsuspend_skill_switch_dialog_yes(skill_switch)
 	orig_MenuCallbackHandler_unsuspend_skill_switch_dialog_yes(self, skill_switch)
-	if EPSS.settings.autobind_skills_2 then
-		local profile_idx = managers.multi_profile and managers.multi_profile._global._current_profile
-		if profile_idx and profile_idx == skill_switch then
-			managers.skilltree:switch_skills(skill_switch)
-			self:refresh_node()
-		end
+
+	if EPSS.settings.autobind_skillsets then
+		managers.multi_profile:epss_bind_skillset(skill_switch)
 	end
 end