+++ Extra Profiles and Skill Sets/lua/skilltreemanager.lua
@@ -1,8 +1,13 @@
---Automatically unlock all skill sets if using autobind
+--Unlock all skill sets if using autobind
 Hooks:PostHook(SkillTreeManager, "load", "EPSS-PostHook-SkillTreeManager:load", function(self)
-	if EPSS.settings.autobind_skills_2 then
-		for i, switch_data in ipairs(self._global.skill_switches) do
+	if not EPSS.settings.autobind_skillsets then
+		return
+	end
+
+	for i, switch_data in ipairs(self._global.skill_switches) do
+		if not switch_data.unlocked then
 			switch_data.unlocked = true
+			switch_data.specialization = Application:digest_value(1, true)
 		end
 	end
 end)